Fixes infinite loop particle bug with blueprint copy/paste
Code analysis suggested this line could be simplified, but it resulted
in an infinite loop call to itself: spawnParticle(particle, dustOptions,
(double) x, (double) y, (double) z);

+++ src/main/java/world/bentobox/bentobox/api/user/User.java
@@ -589,7 +589,7 @@ public class User {
      * @param z Z coordinate of the particle to display.
      */
     public void spawnParticle(Particle particle, Particle.DustOptions dustOptions, int x, int y, int z) {
-        spawnParticle(particle, dustOptions, x, y, z);
+        spawnParticle(particle, dustOptions, (double) x, (double) y, (double) z);
     }
 
     /* (non-Javadoc)
